#  Optimized the JsonWriter APIs and expanded the unit test.