--- a/src/System.Text.Formatting/System/Text/Formatting/Formatters/FormatterThrowHelper.cs
+++ b/src/System.Text.Formatting/System/Text/Formatting/Formatters/FormatterThrowHelper.cs
@@ -1,0 +1,22 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.CompilerServices;
+
+namespace System.Text.Formatting
+{
+    internal static class FormatterThrowHelper
+    {
+        public static void ThrowInvalidOperationException(string message)
+        {
+            throw GetInvalidOperationException(message);
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static InvalidOperationException GetInvalidOperationException(string message)
+        {
+            return new InvalidOperationException(message);
+        }
+    }
+}