Compared to [previous benchmark results](https://github.com/dotnet/corefxlab/pull/2254#issuecomment-386201486), we are 2-4x faster, which makes the writer faster than Newtonsoft.

**Optimizations:**
- Made ArrayFormatter GetSpan by avoiding unnecessary construction of ArrayFormatter and casts.
- Collapsed several method chains to avoid redundant IsUtf8/IsUtf16/etc. checks, and reduced calls to EnsureBuffer/Advance by pre-calculating exactly how many bytes are required.
- Added separate private methods for each encoding (utf-8, utf-16, slow), which resulted in some code duplication.
- Ported over the optimized Utf8Formatter APIs to WriteDigitsUInt64D, to make WriteValue(long) faster, and skip the StandardFormatter switch statement all together. TODO: Port this fix for all numeric overloads (mainly ulong).
- Optimized EnsureBuffer by removing unnecessary checks.
- Added ThrowHelpers to help with inlining.

**Other fixes:**
- Some code formatting clean-up and fix of `var` usage.

``` ini

BenchmarkDotNet=v0.10.14.534-nightly, OS=Windows 10.0.16299.431 (1709/FallCreatorsUpdate/Redstone3)
Intel Xeon CPU E5-1620 v2 3.70GHz, 1 CPU, 8 logical and 4 physical cores
Frequency=3604598 Hz, Resolution=277.4234 ns, Timer=TSC
.NET Core SDK=2.1.300-rtm-008838
  [Host]     : .NET Core 2.1.0-rtm-26514-02 (CoreCLR 4.6.26514.02, CoreFX 4.6.26514.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-rtm-26514-02 (CoreCLR 4.6.26514.02, CoreFX 4.6.26514.02), 64bit RyuJIT


```
|                         Method |         Target | Formatted |        Mean |       Error |      StdDev |      Median |  Gen 0 | Allocated |
|------------------------------- |--------------- |---------- |------------:|------------:|------------:|------------:|-------:|----------:|
|      **WriterSystemTextJsonBasic** | **InvariantUtf16** |     **False** | **14,794.9 ns** | **136.8339 ns** | **114.2625 ns** | **14,760.6 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic | InvariantUtf16 |     False | 17,417.6 ns | 258.7934 ns | 242.0755 ns | 17,504.0 ns | 0.0610 |     416 B |
| WriterSystemTextJsonHelloWorld | InvariantUtf16 |     False |    149.4 ns |   0.2963 ns |   0.2474 ns |    149.4 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld | InvariantUtf16 |     False |    282.2 ns |   4.9845 ns |   4.6625 ns |    279.6 ns | 0.0486 |     256 B |
|      **WriterSystemTextJsonBasic** | **InvariantUtf16** |      **True** | **15,751.3 ns** | **298.0006 ns** | **292.6764 ns** | **15,703.0 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic | InvariantUtf16 |      True | 31,092.0 ns |  72.0731 ns |  42.8896 ns | 31,089.7 ns | 0.0916 |     584 B |
| WriterSystemTextJsonHelloWorld | InvariantUtf16 |      True |    151.3 ns |   2.9176 ns |   2.8654 ns |    153.2 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld | InvariantUtf16 |      True |    384.4 ns |   1.9225 ns |   1.3901 ns |    384.1 ns | 0.0806 |     424 B |
|      **WriterSystemTextJsonBasic** |  **InvariantUtf8** |     **False** | **15,637.4 ns** | **320.2385 ns** | **785.5505 ns** | **15,267.8 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |  InvariantUtf8 |     False | 17,951.5 ns | 335.2059 ns | 313.5518 ns | 17,902.8 ns | 0.0610 |     416 B |
| WriterSystemTextJsonHelloWorld |  InvariantUtf8 |     False |    207.7 ns |   0.6127 ns |   0.4053 ns |    207.6 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |  InvariantUtf8 |     False |    285.1 ns |   4.2451 ns |   3.5449 ns |    286.4 ns | 0.0486 |     256 B |
|      **WriterSystemTextJsonBasic** |  **InvariantUtf8** |      **True** | **16,288.8 ns** | **224.7983 ns** | **210.2764 ns** | **16,222.2 ns** |      **-** |       **0 B** |
|          WriterNewtonsoftBasic |  InvariantUtf8 |      True | 30,375.8 ns | 472.5879 ns | 442.0589 ns | 30,474.8 ns | 0.0916 |     584 B |
| WriterSystemTextJsonHelloWorld |  InvariantUtf8 |      True |    214.5 ns |   4.5213 ns |   4.6430 ns |    212.8 ns |      - |       0 B |
|     WriterNewtonsoftHelloWorld |  InvariantUtf8 |      True |    386.6 ns |   0.8402 ns |   0.7016 ns |    386.5 ns | 0.0806 |     424 B |